requirements.txt
Telethon==1.19.1
python-dotenv==0.15.0
requests~=2.25.1